// propertyApp/src/app/protected/interfaces/izconfig.ts
export interface IZConfig {
    id: string;
    accountid: string;
    currencyid: string;
    taxid: string;
    smtpfromname: string;
    smtpfromemail: string;
    smtpfrompassword: string;
    smtptoname: string;
    smtptoemail: string;
    smtptomensaje: string;
    created_by?: string;
    created_at?: string;
    modified_by?: string;
    modified_at?: string;
}